[FIX] Fix all Endpoints, Update Error and valid messages, Fix all non working Tests

+++ api/app/views/meals.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import (
     JWTManager, jwt_required, create_access_token,
     get_jwt_identity
 )
-from app.views.auth.validate import require_admin,mealname_and__menuitem_validator
+from app.views.auth.validate import *
 
 
 class MealOptions(Resource):
@@ -19,37 +19,47 @@ class MealOptions(Resource):
         """
         Method gets a meal by id.
         """
+        active_user = User.query.filter_by(email=get_jwt_identity()).first()
         meal = Meal.query.filter_by(id=id).first()
         if meal is None:
             return {"status":"Failed!!",
-            "message":"Meal id does not exist.Please enter a valid meal id"}
-        
-        response = meal.json_dump()
-        return response
+            "message":"Meal id does not exist.Please enter a valid meal id"},404
+        if meal.user_id == active_user:
+            response = meal.json_dump()
+            return response
+        return {'message':'Meal does not exist'},404
+
     @jwt_required
     @require_admin
     def put(self, id):
         """
-        Method updates meal.
+        Method updates that Updates the meal of certain ID.
         """
         json_data = request.get_json(force=True)
+
+        if 'meal_name' not in json_data or 'price' not in json_data:
+            return {"status": "Failure", "data": "Please provide a meal a name or price."}
+        
         meal_name = json_data['meal_name']
         meal_price = json_data['price']
+        active_user = User.query.filter_by(email=get_jwt_identity()).first()
         meal = Meal.query.filter_by(id=id).first()
-        if 'meal_name' not in json_data or 'price' not in json_data:
-            return {"status":"Failure","data":"Please provide a meal a name or price."}
         if meal is None:
             return {"status":"Failure",
-            "message":"Meal id does not exist.Please enter a valid meal id"}
-        if meal_name == '' or not mealname_and__menuitem_validator(json_data['meal_name']) or price == '':
-            return {"status":"Failed!!",
-            "message":"Meal name can not be empty.Please enter a valid meal details"}
-        else:
-            meal.meal_name = meal_name
-            meal.meal_price = meal_price
-            db.session.commit()
-            response = meal.json_dump()
-            return{"status": "success", "data": response}, 200
+            "message":"Meal id does not exist.Please enter a valid meal id"},404
+        
+        if meal.user_id == current_user.id:
+            if meal_name == '' or not mealname_and__menuitem_validator(json_data['meal_name']) or price == '':
+                return {"status":"Failed!!",
+                "message":"Meal name can not be empty.Please enter a valid meal details"}
+            else:
+                meal.meal_name = meal_name
+                meal.meal_price = meal_price
+                db.session.commit()
+                response = meal.json_dump()
+                return{"status": "success", "data": response}, 200
+        return {'message':'meal Does not exit'},404
+
     @jwt_required
     @require_admin
     def delete(self, id):
@@ -57,15 +67,17 @@ class MealOptions(Resource):
          Method deletes a meal by id.
         """
         json_data = request.get_json(force=True)
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         meal= Meal.query.filter_by(id=id).first()
         if  meal is None:
             return {"status":"Failed!!",
             "message":"Meal id does not exist.Please enter a valid meal id"}
-        else:
+        if meal.user_id == current_user.id:
             Meal.query.filter_by(id=id).delete()
             db.session.commit()
             response = json.loads(json.dumps(json_data))
             return {"status": "deleted!", "data": response}, 200
+        return {'Error':'Meal Does not exit'},404
 
 class MealLists(Resource):
     """
@@ -78,9 +90,13 @@ class MealLists(Resource):
         """
         Getting Meals Lists
         """
-        meals = Meal.query.all()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        meals = Meal.query.filter_by(user_id=current_user.id)
+        if meals is None:
+            return{"message": "No Meal List At The Moment"}
         response = [meal.json_dump() for meal in meals]
-        return {"status": "success", "data": response}, 200
+        return {"status": "success", "message": response}, 200
+
     @jwt_required
     @require_admin
     def post(self):
@@ -88,14 +104,19 @@ class MealLists(Resource):
          Posting Meals
         """
         json_data = request.get_json(force=True)
-        meal_name = json_data['meal_name']
-        meal_price = json_data['price']
+
         if 'meal_name' not in json_data or 'price' not in json_data:
-            return {"status":"Failed!","data":"Please provide a meal a Details."}
-        if meal_name == '' or meal_price =='':
+            return {"status": "Failed!", "data": "Please provide a meal_name and price Details."},406
+
+        meal_name = space_stripper(json_data['meal_name'])
+        meal_price = json_data['price']
+        if meal_name == '' or not name_validator(meal_name) or meal_price == '':
             return {"status":"Failed",
-            "message":"Meal name can not be empty.Please enter a valid meal Details"}
-        meal = Meal(meal_name=meal_name, meal_price = meal_price)
+            "message":"Meal name can not be empty.Please enter a valid meal Details"},406
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        user_id = current_user.id
+        meal = Meal(meal_name=meal_name, user_id=user_id,
+                    meal_price=meal_price)
         meal.save()
         response = json.loads(json.dumps(meal.json_dump()))
         return {"status": "success", "data": response}, 201
