+++ api/app/views/menu.py
@@ -1,11 +1,13 @@
 from flask import json, request
 from flask_restful import Resource,reqparse
-from app.models import db, Menu,Meal
+from app.models import db, Menu,Meal, User
 from flask_jwt_extended import (
     JWTManager, jwt_required, create_access_token,
     get_jwt_identity
 )
-from  app.views.auth.validate import require_admin
+from  app.views.auth.validate import require_admin, space_stripper, name_validator, num_check
+
+
 class MenuOptions(Resource):
     """
     Update,Delete,PUT and migrate
@@ -16,12 +18,16 @@ class MenuOptions(Resource):
         """
         Method gets a menu by id.
         """
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         menu = Menu.query.filter_by(id=id).first()
         if menu is None:
             return {"status":"Failed!!",
             "message":"Menu id does not exist.Please enter a valid menu id"}
-        response = menu.json_dump()
-        return response
+        if menu.user_id == current_user.id:
+            response = menu.json_dump()
+            return response
+        return {"status": "Failed!", "message": "Meal does not exist."}, 404
+
     @jwt_required
     @require_admin
     def put(self, id):
@@ -30,26 +36,31 @@ class MenuOptions(Resource):
         """
         json_data = request.get_json(force=True)
         menu = Menu.query.filter_by(id=id).first()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        meal = Meal.query.filter_by(id=json_data['meal_id']).first()
         if 'meal_id' not in json_data or 'menu_item' not in json_data:
             return {"status":"Failed!","message":"Please provide a meal_id and menu_item to update."}
         if  menu is None:
             return {"status":"Failed!!",
             "message":"Menu id does not exist.Please enter a valid meal id"}
-        meal= Meal.query.filter_by(id=json_data['meal_id']).first()
         if  meal is None:
             return {"status":"Failed!!",
             "message":"Meal id does not exist.Please enter a valid meal id"}
-        if json_data['menu_item']== '':
+        if not num_check(json_data['meal_id']):
+            return {"status": "Failed!",
+                   "message": "Meal id must be an integer"}, 406
+        if json_data['meal_item'] == '' or json_data['meal_id'] == '':
             return {"status":"Failed!!",
             "message":"Menu item can not be empty.Please enter a valid menu item"}
-        else:
-            if json_data['meal_id']:
+        if meal.user_id == current_user.id:
+            if menu.user_id == current_user.id:
                 menu.meal_id = json_data['meal_id']
-            if json_data['menu_item']:
-                menu.menu_item = json_data['menu_item']
-            db.session.commit()
-            response = menu.json_dump()
-            return{"status": "success", "data": response}, 200
+                menu.meal_item = space_stripper(json_data['meal_item'])
+                db.session.commit()
+                response = menu.json_dump()
+                return{"status": "success", "data": response}, 200
+            return {"message": "Item does not exist!"}, 404
+        return {"message": "Meal does not exist!"}, 404
 
     @jwt_required
     @require_admin
@@ -59,14 +70,16 @@ class MenuOptions(Resource):
         """
         json_data = request.get_json(force=True)
         menu= Menu.query.filter_by(id=id).first()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         if  menu is None:
             return {"status":"Failed!!",
             "message":"Menu id does not exist.Please enter a valid meal id"}
-        else:
+        if menu.user_id == current_user.id:
             Menu.query.filter_by(id=id).delete()
             db.session.commit()
             response = json.loads(json.dumps(json_data))
             return {"status": "deleted!", "data": response}, 200
+        return {"message": "Meal does not exist."}, 404
 
 
 class MenuList(Resource):
@@ -78,9 +91,13 @@ class MenuList(Resource):
         """
         Method to get all menus.
         """
-        menus = Menu.query.all()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        menus = Menu.query.filter_by(user_id=current_user.id)
+        if menus is None:
+            return{"message": "You do not have menus at the moment."}, 404
         response = [menu.json_dump() for menu in menus]
         return {"status": "success", "data": response}, 200
+
     @jwt_required
     @require_admin
     def post(self):
@@ -93,24 +110,23 @@ class MenuList(Resource):
             return {"status":"Failed!",
                     "Message":"Please a valid provide a meal_id and menu_cartegory in [Breakfast,Beverages,Lunch,,Super] to add to the menu."}
        
-        # menu_cart = ['Breakfast', 'Beverages', 'Lunch', 'Super']
-        # check_cart = [u for u in menu_cart if not u]
-        
-        # if not check_cart:
-        #     return {'status':"Failure", 'message':'Enter {}'.format(menu_cart)}
         
         menu_item = json_data['menu_cartegory']
         meal_id   = json_data['meal_id']
         meal= Meal.query.filter_by(id=meal_id).first()
-        if menu_item == '':
+        if space_stripper(menu_item) == '':
             return {"status":"Failed",
-            "message":"Meal name can not be empty.Please enter a valid menu item"}
-        elif meal_id == '' :
+            "message":"Meal name can not be empty.Please enter a valid menu item"},406
+        if meal_id == '' :
             return {"status":"Failed",
-            "message":"Meal id can not be empty.Please enter a valid meal id"}
+            "message":"Meal id can not be empty.Please enter a valid meal id"},406
         if  meal is None:
             return {"status":"Failed!!","message":"Please enter a valid meal id"}
-        menu = Menu(meal_id=meal_id,menu_item=menu_item)
-        menu.save()
-        response = json.loads(json.dumps(menu.json_dump()))
-        return {"status": "success", "data": response}, 201
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        user_id = current_user.id
+        if meal.user_id == current_user.id:
+            menu = Menu(meal_id=meal_id, meal_item=meal_item, user_id=user_id)
+            menu.save()
+            response = json.loads(json.dumps(menu.json_dump()))
+            return {"status": "success", "data": response}, 201
+        return {"message": "Meal does not exist"}, 404
