+++ api/app/views/orders.py
@@ -6,7 +6,7 @@ from flask_jwt_extended import (
     JWTManager, jwt_required, create_access_token,
     get_jwt_identity
 )
-from app.views.auth.validate import require_admin
+from app.views.auth.validate import require_admin, num_check, name_validator
 
 class Orders(Resource):
     """
@@ -18,41 +18,44 @@ class Orders(Resource):
         """
         Method gets a order by id.
         """
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         order = Order.query.filter_by(id=id).first()
         if order is None:
             return {"status":"Failed!!",
             "message":"Order id does not exist.Please enter a valid order id"}
-        response = order.json_dump()
-        return response
+        if order.user_id == current_user.id:
+            response = order.json_dump()
+            return response
+        return {"message": "Order does not exist"}, 404
+
     @jwt_required
     def put(self, id):
         """
         Method updates order.
         """
         json_data = request.get_json(force=True)
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        item = Menu.query.filter_by(id=json_data['item_id']).first()
         order = Order.query.filter_by(id=id).first()
         if  order is None:
             return {"status":"Failed!!",
-            "message":"Order id does not exist.Please enter a valid order id"}
-        if 'user_id' not in json_data or \
-             'item_id' not in json_data or 'quantity' not in json_data:
-              return {"status": "Failed!",
-               "message": "Please supply user id, item id and quantity"},406
-        item= Menu.query.filter_by(id=json_data['item_id']).first()
+            "message":"Order id does not exist.Please enter a valid order id"},404
+
+        if not num_check(json_data['item_id']) or not num_check(json_data['quantity']):
+              return {"status": "Failed!", "meassage": "Meal_id,item_id and quantity must be integers."}, 406
+
         if  item is None:
             return {"status":"Failed!!",
-            "message":"Item id does not exist.Please enter a valid Item id"}
-        user = User.query.filter_by(id=json_data['user_id']).first()
-        if  user is None:
-            return {"status":"Failed!!",
-            "message":"User id does not exist.Please enter a valid User id"}
-        else:
-            order.user_id = json_data['user_id']
-            order.item_id = json_data['item_id']
-            order.quantity = json_data['quantity']
-            db.session.commit()
-            response = order.json_dump()
-            return{"status": "success", "data": response}, 200
+            "message":"Item id does not exist.Please enter a valid Item id"},404
+        if item.user_id == current_user.id:
+            if order.user_id == current_user.id:
+                order.item_id = json_data['item_id']
+                order.quantity = json_data['quantity']
+                db.session.commit()
+                response = order.json_dump()
+                return{"status": "success", "data": response}, 200
+            return {"message": "Order does not exist"}, 404
+        return {"message": "Item does not exist!"}, 404
 
     @jwt_required
     def delete(self, id):
@@ -61,15 +64,16 @@ class Orders(Resource):
         """
         json_data = request.get_json(force=True)
         order= Order.query.filter_by(id=id).first()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
         if  order is None:
             return {"status":"Failed!!",
             "message":"Order id does not exist.Please enter a valid order id"}
-        else:
+        if order.user_id == current_user.id:
             Order.query.filter_by(id=id).delete()
             db.session.commit()
             response = json.loads(json.dumps(json_data))
             return {"status": "deleted!", "data": response}, 200
-
+        return {"message": "Order does not exist."}, 404
 
 class OrderLists(Resource):
     """
@@ -81,41 +85,41 @@ class OrderLists(Resource):
         """
         Method to get all orders.
         """
-        orders = Order.query.all()
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        orders = Order.query.filter_by(user_id=current_user.id)
+        if orders is None:
+            return{"message": "You do not have orders at the moment."}, 404
         response = [order.json_dump() for order in orders]
         return {"status": "success", "message": response}, 200
+
     @jwt_required
     def post(self):
         """
          Method creates an order.
         """
         json_data = request.get_json(force=True)
-        if 'user_id' not in json_data or \
-             'item_id' not in json_data or 'quantity' not in json_data:
-              return {"status": "Failed!",
-               "message": "Please supply user id, item id and quantity"},406
+        if 'item_id' not in json_data or 'quantity' not in json_data:
+            return {"status": "Failed!",
+                      "message": "Please supply user id, item id and quantity"}, 406
+        if not num_check(json_data['item_id']) or not num_check(json_data['quantity']):
+            return {"status": "Failed!", "message": "Item and quantity id must be integers"}, 406
+        
         item_id = json_data['item_id']
-        user_id   = json_data['user_id']
         quantity = json_data['quantity']
         item= Menu.query.filter_by(id=item_id).first()
-        user= User.query.filter_by(id=user_id).first()
         if item_id == '':
             return {"status":"Failed",
-            "message":"Item id can not be empty.Please enter a valid item id"}
-        elif user_id == '' :
-            return {"status":"Failed",
-            "message":"User id can not be empty.Please enter a valid user id"}
+            "message":"Item id can not be empty.Please enter a valid item id"},406
         elif quantity == '' :
             return {"status":"Failed",
-            "message":"Quantity can not be empty.Please enter  valid quantity"}
+            "message":"Quantity can not be empty.Please enter  valid quantity"},406
         if  item is None:
             return {"status":"Failed!!",
-            "message":"Item id does not exist.Please enter a valid item id"}
-        if  user is None:
-            return {"status":"Failed!!",
-            "message":"User id does not exist.Please enter a valid user id"}
-        else:
-            order = Order(user_id=user_id,item_id=item_id,quantity=quantity)
-            order.save()
-            response = json.loads(json.dumps(order.json_dump()))
-            return {"status": "success", "data": response}, 201
+            "message":"Item id does not exist.Please enter a valid item id"},404
+            
+        current_user = User.query.filter_by(email=get_jwt_identity()).first()
+        user_id = current_user.id
+        order = Order(user_id=user_id, item_id=item_id, quantity=quantity)
+        order.save()
+        response = json.loads(json.dumps(order.json_dump()))
+        return {"status": "success", "data": response}, 201
